Add TLS support to the ctl gateway client
